xaie: Enable backend selection at init time
All IO backends are compiled no matter the configuration. The only
thing that changes is the XAIE_DEFAULT_BACKEND macro that points to
the correct backend as specified by a compile-time flag. There is no
reason why the backend selection cannot be made at runtime, giving
more control to the application as to how it wants to use the
library.

Add a configuration item specifying the backend. When the library is
initialized, the backend will be chosen based on this parameter. If
the user-selected backend index is out of range, the library will
default to the LIBMETAL backend.

Signed-off-by: Joel Nider <joel.nider@amd.com>